+++ roseau/load_flow/models/tests/test_lines.py
@@ -2,8 +2,8 @@ import numpy as np
 import pytest
 from pint import DimensionalityError
 
-from roseau.load_flow import RoseauLoadFlowException, RoseauLoadFlowExceptionCode
-from roseau.load_flow.models import Bus, Line, LineParameters
+from roseau.load_flow.exceptions import RoseauLoadFlowException, RoseauLoadFlowExceptionCode
+from roseau.load_flow.models import Bus, Ground, Line, LineParameters
 from roseau.load_flow.units import Q_
 
 
@@ -61,3 +61,37 @@ def test_lines_units():
     line = Line("line", bus1=bus1, bus2=bus2, parameters=lp, length=5)
     with pytest.raises(DimensionalityError, match=r"Cannot convert from 'ampere' \(\[current\]\) to 'km'"):
         line.length = Q_(6.5, "A")
+
+
+def test_line_parameters_shortcut():
+    bus1 = Bus("bus1", phases="abcn")
+    bus2 = Bus("bus1", phases="abcn")
+
+    #
+    # Without shunt
+    #
+    lp = LineParameters("lp", z_line=np.eye(4, dtype=complex))
+
+    # Z
+    line = Line("line", bus1=bus1, bus2=bus2, parameters=lp, length=Q_(50, "m"))
+    assert np.allclose(line.z_line.m_as("ohm"), 0.05 * np.eye(4, dtype=complex))
+
+    # Y
+    assert not line.with_shunt
+    assert np.allclose(line.y_shunt.m_as("S"), np.zeros(shape=(4, 4), dtype=complex))
+
+    #
+    # With shunt
+    #
+    z_line = 0.01 * np.eye(4, dtype=complex)
+    y_shunt = 1e-5 * np.eye(4, dtype=complex)
+    lp = LineParameters("lp", z_line=z_line, y_shunt=y_shunt)
+
+    # Z
+    ground = Ground("ground")
+    line = Line("line", bus1=bus1, bus2=bus2, parameters=lp, length=Q_(50, "m"), ground=ground)
+    assert np.allclose(line.z_line.m_as("ohm"), 0.05 * z_line)
+
+    # Y
+    assert line.with_shunt
+    assert np.allclose(line.y_shunt.m_as("S"), 0.05 * y_shunt)